pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "automosaic"
version = "1.0.0"
description = "画像・動画の顔検出自動モザイク処理ツール"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "AutoMosaic Contributors"},
]
keywords = ["face-detection", "mosaic", "privacy", "cli", "opencv"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "opencv-python>=4.8.0",
    "mediapipe>=0.10.0",
    "click>=8.1.0",
    "tqdm>=4.65.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
]

[project.optional-dependencies]
yolo = ["ultralytics>=8.0.0"]
whitelist = ["face_recognition>=1.3.0"]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
all = ["automosaic[yolo,whitelist,dev]"]

[project.scripts]
automosaic = "automosaic.cli:main"

[project.urls]
Homepage = "https://github.com/automosaic/automosaic"
Repository = "https://github.com/automosaic/automosaic"
Issues = "https://github.com/automosaic/automosaic/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=automosaic --cov-report=term-missing"
